attempted modified tests
