+++ api/server.spec.js
@@ -1,7 +1,13 @@
 const server = require('./server');
 const request = require('supertest');
 
+const db = require('../database/db-config');
+
 describe('request to server', () => {
+  beforeAll(async () => {
+    return await db.seed.run();
+  });
+
   it('responds with 200', async done => {
     await request(server)
       .get('/')
